\name{test}
\alias{test}
\docType{data}
\title{
Example of a data frame
}
\description{
a dataframe containing 200 rows (i.e. 200 individuals). The first column contains the name of the population ("A", "B", "C", "D" and "E") to which the individual belongs, the 11 others contain quantitative measures.
}
\usage{data("test")}

\references{

Michalski S.G., Durka W. 2015. Data from: Separation in flowering time contributes to the maintenance of sympatric cryptic plant lineages. Ecology and Evolution 5 (11): 2172-2184. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.bk5hk.
}

\examples{

data(test)

}
